mesh: propagate supervision events to the proc mesh (#42)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/42

This wires up the supervision mechanism to deliver supervision events to the proc mesh itself. Currently we interpret these as a proc crash.

This will then allow us to install a default policy to always fail on such errors (but, crucially, never silently). We can then drill into fine-grained delivery of the supervision events to the actor mesh level instead, from which they can be recovered.

Reviewed By: shayne-fletcher

Differential Revision: D74830956

fbshipit-source-id: 21ac2ff072c33ab47880a9dd510c11b3baa5618c