Add WatermarkAccumulator (#60)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/60

We need an accumulator to track the lowest value among ranks. In order to distinguish from regular min accumulator, we call it low watermark accumulator. This diff introduces this accumulator.

This implementation does not do any optimization, and uses a HashMap to store the states. The main focus of this diff is to get the API correct. We will do the optimization in followups.

Reviewed By: mariusae

Differential Revision: D75166323

fbshipit-source-id: 1019c40e9a140f0aa4b710468e0cd654ca4df63e