Start simulator lazily (#150)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/150

The old pattern required us to start the simnet and proxy before we provided a proxy address, with us adding the proxy address every time we dialed a sim address. If we use a OnceLock, we can start the simnet once when we bootstrap the simulator and start the proxy at the proxy address.

We would also temporarily store the operational message receiver as a field in the simnet handle and take this away when we bootstrapped. Now we can just directly return the operational message receiver when we start the simnet during our bootstrap

Reviewed By: kaiyuan-li

Differential Revision: D75900567

fbshipit-source-id: 306d0c37adc1759019cd0e4e69050470b1e2751e