make tensor_engine a crate feature of monarch_extension (#194)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/194

Allow for conditional compilation of `tensor_engine`-related code.

This is a pre-cursor to a more substantive separation in layering of these two components.

In the end-state, `hyperactor` and `actor_mesh` shouldn't know about the tensor-engine (or torch) at all. This will help with testing, make it eeasier on environments without CUDA, etc.

This is just a convenient way to start, because it means we can turn the feature off and run Python tests to prove to ourselves there is actually no dependency, before we start a disruptive code migration.
ghstack-source-id: 289017940
exported-using-ghexport

Reviewed By: vidhyav

Differential Revision: D76232031

fbshipit-source-id: fa9a9db09cc8e2357a37848ca1664a9770a1f55e