hello tensor engine (#188)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/188

Add the initialize controller class `mesh_controller` that can implement the tensor engine on top of the ProcMesh/ProcActor API.

The only example is currently a "hello world" that allocates and fetches a tensor.

Follow up PRs will integrate creating meshes this way into the testing code more deeply and fix the issues that come up with it.

This design assumes that supervision of stuff is going to be handled by the actor system and that tensor compute can just rely on that for monitoring and stuckness detection stuff.

This design has no ClientActor, and the ControllerActor only exists as a Instance handle for reading messages from the workers (which send some controller messages).

This does not attempt to clean up the existing RustController system yet, since it isn't feature equivalent or tested with it.
ghstack-source-id: 289169944
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D75909313

fbshipit-source-id: 2e25a6699e03a10b9ac2dc1ffa57d11759427157