Avoid repeat error reports (#218)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/218

If an op depends on two nodes that are failing at the time it is issued, then history will generate two failure messages for the sequence number. If there was a future attached to the sequence number, then the second failre will cause a KeyError in the invocation dicationary since the first instance already removed it.

This depends on a race condition to hit because the messages for the failure must already be present when the node failing is being added.
ghstack-source-id: 289532983
exported-using-ghexport

Reviewed By: suo

Differential Revision: D76372510

fbshipit-source-id: a0c60dd70c3a22f8dfb0de529493644e4baadc42