mesh: fix sub-mesh message ordering (#220)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/220

Sub-set the selection to the selection that represents the mesh's
view of the root mesh. We need to do this because the comm actor uses the slice as the stream key; thus different sub-slices will result in potentially out of order delivery.

The real fix is to introduce an explicit stream key, associated with the root mesh. This will be done in a follow-up diff.

Reviewed By: shayne-fletcher

Differential Revision: D76372021

fbshipit-source-id: aece4b383df31ca5c3008d268c4d295b084491d7