ProcMesh activate() / mock slicing (#264)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/264

See test_python_actor.py for an example.

DeviceMesh is hidden inside of ProcMesh wrapper, so that you can use ProcMesh as the way to tell the tensor engine what to do.

Currently we cannot spawn sliced proc meshes so we have to fake the slicing behavior for ProcMesh to allow it to stil be used as a sliced mesh for the purpose of the tensor engine.
ghstack-source-id: 290789083
exported-using-ghexport

Reviewed By: suo

Differential Revision: D76556227

fbshipit-source-id: 9df0d6e2114d6da2f1821c178248cade4314020f