(2/n torchx-allocator)(monarch/tools) implement a way to resolve hostname to ipv6 or ipv4 address (#295)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/295

implement `monarch.tools.network.get_ip_addr(hostname)` that resolves a hostname to an ip address by:

1. Look up a TCP (`SOCK_STREAM`) compatible ipv6 address
2. If not found, fall-back to TCP compatible ipv4 address
3. Error if no ipv6 or ipv4

This is required since hyperactor's `ChannelAddr` for the `tcp` transport (e.g. `tcp!slurm-compute-node-0:26600`) takes a socketaddr (ip + port) not a hostname. So the caller has to resolve the hostname (typically queried from the scheduler) to an ip.

Reviewed By: ahmadsharif1

Differential Revision: D76846286

fbshipit-source-id: 47f81f87cec4351bcd03a27c9e4b977689565a6a