(3/n torchx-allocator)(monarch/tools) add commands.server_ready function and hostnames to mesh_spec (#296)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/296

TorchX's `status` API returns a struct that has `replica.hostname` field. However it is not always filled for all schedulers. https://github.com/pytorch/torchx/pull/1080 makes it such that the slurm scheduler in TorchX fills out the hostname information.

This PR adds a `hostnames` field to `monarch.tools.mesh_sepc.MeshSpec` and fills it up with the hostnames returned by TorchX.

This information will be used in PR (5/n) to implement a `TorchXAllocator`

Reviewed By: suo

Differential Revision: D76847192

fbshipit-source-id: 4d55083009ef9dd6ed46717fd375f5a49ee86a95