Expose simulator sleep to python (#292)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/292

We expose the SleepEvent from the hyperactor runtime simulator into python. It is non-blocking. It can be used in Python actors to simulate latency. Once it starts, it will not allow code execution to progress until the DES event loop has advanced to the time the sleep is completed and sends a signal to wake it up.

Explanation for changes in simnet.rs:
- We re-introduce a debounce period where we do not advance to the next time with events. This was originally introduced to speed up the tensor engine use case which was deprioritized. By having a debounce, we will prevent the event loop from over advancing
- We set the default training_script_state to TrainingScriptState::Waiting in order to allow for run-ahead. For the tensor engine use case, we wanted to never advance past the wall time of when client messages came in, until we hit a fetch_shard().result() in which case it was safe to advance past wall time.

Reviewed By: benjipelletier

Differential Revision: D76765388

fbshipit-source-id: 03017fb5a95d9aa0773cb2d375e3282313b46076