move History into mesh_controller in prep for major refactor (#277)
Summary:
mesh_controller will respond directly to PortId when things are finished inside controller. This is a big enough change that I am splitting the code from the other controller. This is just a wholesale copy of the history code, deleting the parts we don't use in this file. The next diff makes the changes needed for mesh_controller.

Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/277
ghstack-source-id: 291379179

Reviewed By: suo

Differential Revision: D76643800

fbshipit-source-id: 4a95546bea4b22edf1ce601d884511a9740f261e