Get history ready for working with ports (#278)
Summary:
This cleans up the history object in order to replace the message returns with direct sending to the return port.

In particular, it clarifies the lifetimes of Invocation objects an Ref s by replacing the hash tables with Arc links. Now invocations live as long as there is a Ref referring to them, or if it is still in flight. Status enum make sure we do not hold users longer than needed.

Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/278
ghstack-source-id: 291379176

Reviewed By: suo

Differential Revision: D76649221

fbshipit-source-id: 2992938be1f8f59ccd18559d3a09de151d5f7ca3