Make timed_test robust to deadlocked tokio runtimes (#347)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/347

Make it so that the timeout is guaranteed to fire, even if the underlying runtime is completely blocked on some bogus work that never awaits.

This type of code should *never* be written; all native code that may deadlock should be run on an isolated thread via `spawn_blocking` or similar. However it's important to make our tests robust to it.
ghstack-source-id: 292774516

Reviewed By: mariusae

Differential Revision: D77342496

fbshipit-source-id: c27a5883b3f83568082146f4ae023836dc291f43