general in-proc ActorRef -> ActorHandle downcasting (#350)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/350

Provide a general purpose mechanism for downcasting any local (in-proc) ActorRef<A> to its corresponding ActorHandle<A>. This allows these references to be coordinated externally, but for the actors to communicate locally (including using handlers that aren't exported, e.g., because they may not contain serializable messages).

The intent is to use this in the tensor engine, to allow the controller to instruct actors to coordinate locally, passing objects like tensor references directly.

This required building a general instance registration mechanism, which we can use to replace the ledger in a follow-up change.
ghstack-source-id: 293003584

Reviewed By: shayne-fletcher

Differential Revision: D77377382

fbshipit-source-id: 7f540e3a8cef38717e89543d6152a2e37f0445d8