cleanup (#356)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/356

In several places, we have a pattern of newtyping an Arc<State>, to provide a publicly accessible objects (e.g., Mailbox, Proc, etc.) that is cloneable.

This is just a small cleanup to standardize this pattern into

```
struct Outer {
  inner: Arc<Inner>,
}

...
```

Possibly we should also standardize on calling such objects "Handles", since they are not owned, but rather provide a handle to some other object.
ghstack-source-id: 293003582

Reviewed By: suo, shayne-fletcher

Differential Revision: D77395146

fbshipit-source-id: 1cb2cce95ea981da4fa93e5caee48233b79d4af9