Plumb actor supervision events into actor mesh (#280)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/280

This diff pipes actor supervision events
received in the proc mesh (for now) into the respective actor
meshes. In order to do that, we have the following:

1) ActorMesh::next() which should be used to observe lifecycle events of
the actor and handle them. (next diff does the actual lifecycle
observer running in the background) and the handling logic.
2) ProcMesh should only react to unhandled supervision events
that should appear as exactly that.

Reviewed By: moonli

Differential Revision: D76734454

fbshipit-source-id: 0a044c5e096ab43cd8f9e6413bfa197c09f4b91e