python ProcMesh supervision API (#333)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/333

This diff adds supervision API to python ProcMesh class. Proc supervision events will be exposed as an async iterator.

The default behavior of supervision error handling is terminating the client process, this diff provides an API to override this behavior. User can choose to observe the supervision events on a ProcMesh via the new "monitor()" API. When observing starts, supervision error won't automatically exit the client process, instead it pass the error handling responsibility to user who is actively observing the errors.

Reviewed By: colin2328

Differential Revision: D77033984

fbshipit-source-id: 2ffc58f9501a39f50a63eeb4b92fb614af8a8215