mailbox: refactor undeliverable sender lookup (#367)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/367

ghstack-source-id: 293105751
exported-using-ghexport

in D76926674, `lookup_sender` was introduced to let the `CanSend for Mailbox` implementation recover a `PortHandle<Undeliverable<MessageEnvelope>>` if the mailbox had one bound.

this diff refactors that slightly: introduces `bound_return_handle()` for this purpose, and makes both `lookup_sender` and `UnboundedPortSender` private. updates `CanSend::post` to use the new method.

prep for changes coming later in the stack.

Reviewed By: mariusae

Differential Revision: D77449072

fbshipit-source-id: 0c3bb0a702437ad9884f6e1e2d52cd4685b39e21