Enable pyo3 multiple-pymethods feature
Summary:
Currently in third-party/rust/BUCK, the "multiple-pymethods" feature for the `pyo3` crate is always enabled:

https://www.internalfb.com/code/fbsource/[9cf30cd592679bb02bcdcd3d02cb0c319c6f558c]/third-party/rust/BUCK?lines=133268-133269%2C133456%2C133472%2C133477%2C133548

This happens because `py-polars` enables it.

https://www.internalfb.com/code/fbsource/[9cf30cd592679bb02bcdcd3d02cb0c319c6f558c]/third-party/rust/vendor/py-polars-1.29.0/Cargo.toml?lines=125-127%2C131-132

However, third-party/rust/Cargo.toml only depends on `py-polars` on a subset of platforms, not including Windows.

https://www.internalfb.com/code/fbsource/[9cf30cd592679bb02bcdcd3d02cb0c319c6f558c]/third-party/rust/Cargo.toml?lines=1311-1312%2C1340

Under the old Reindeer feature resolver, the "multiple-pymethods" feature gets applied to pyo3 on all platforms anyway, because feature resolution only takes into account x86_64 Linux.

But the new feature resolver (the stack containing D77062707) does an independent feature resolution per platform, and would correctly observe that "multiple-pymethods" is not enabled on Windows.

But existing code in fbcode assumes always being able to rely on multiple-pymethods, including on Windows. This diff declares that.

Reviewed By: diliop

Differential Revision: D77476538

fbshipit-source-id: cc606c1c19e65851c44197298bd1a20bf5a35cbc