Change PythonMessage's signature (#402)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/402

`&[u8]` is much faster than `Vec<u8>` for some reason (See V1 of this diff for comparision). *probably* because:

`Vec<u8>` is mapped to pyo3's `PySequence`, and then it does this under the hood: https://fburl.com/code/6cw3osxf

`&[u8]` is just a pointer to rust, then rust does `memcpy` under the hood, which is much faster.

Reviewed By: moonli, shayne-fletcher

Differential Revision: D77594166

fbshipit-source-id: f296c32ae0539e59391f142b8fb4b236f0dfedc7