remove our custom signal handlers (#385)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/385

These need to be written quite carefully as they need to be async-signal-safe (no log, no printf, no malloc). Let's just not even try, and instead rely on the standard fb implementation which has been worked over forever.

Normally `#[fbinit]` will do this on the main entry point, but because bootstrap_main is a Python entry point that we spawn via the ProcessAllocator we need to call `initFacebook` there as well.
ghstack-source-id: 293873376
exported-using-ghexport

Reviewed By: vidhyav, mariusae

Differential Revision: D77565855

fbshipit-source-id: c5a5f6a3ec180879e4df3544303dcbbcb0ba2756