mesh: key streams off of sender only (#397)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/397

Each stream was previously associated with `(sender, <shape of mesh>)`, the shape being a stand-in for the mesh itself. In practice, each proc is reachable through exactly one mesh, so this is unnecessary.

This change simplifies this to just the sender.

In the future, we'll likely want to support multiple sessions per sender (e.g., for persistent meshes), but that is a straightforward addition we can make as needed.
ghstack-source-id: 293881496
exported-using-ghexport

Reviewed By: shayne-fletcher, vidhyav

Differential Revision: D77609169

fbshipit-source-id: 4babc95eba0d4123f9308b43b5bf37a9dfab2747