Support explicitly killing procs (#408)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/408

Not super graceful yet, but adds some initial bookkeeping to support explicitly
killing a proc mesh via reclaiming the alloc we put in `ProcEvents` and package
in the `Keepalive` token.

In particular, this adds `PyProcMesh.stop` method which consumes the keepalive
token -- leaving the `PyProcMesh` in an effectively unusable state -- and uses it
to stop the alloc.

Reviewed By: vidhyav, mariusae

Differential Revision: D77250211

fbshipit-source-id: 16b30c55cd558c9c52a188aa3364af9379bb6323