heartbeat in child processes to ensure cleanup (#435)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/435

In D77392241, I removed the signal-based cleanup mechanism. The assumption there was that the `rx.recv()` in `bootstrap.rs` would throw an error if the other side hung up, and the error would bubble up and eventually abort the child process.

This assumption is wrong; `rx.recv()` has server-like semantics, not channel-like (which makes sense; there could be many senders, so any individual one disappearing should not abort the receiver!).

But as a result, we were not cleaning up properly if the parent process exited, child processes would just hang around forever. The test added in D77348271 *happened* to pass, because we sent `SIGKILL` to the parent process, triggering unclean shutdown which *will* cause a error on the receiver side. However, a *graceful* shutdown (e.g. from an uncaught Python exception) will not.

A quick solution is to add a simple heartbeat task and kill the process if it fails.
ghstack-source-id: 294670138
exported-using-ghexport

Reviewed By: ahmadsharif1

Differential Revision: D77802426

fbshipit-source-id: 0b7f7b76aa528a22b3921d897c9b04d20914cc69