macros: support generic type parameters in `Named` derive macro, separate `Named` from diagnostic registration (#460)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/460

This change does two things:

1) Implement support for type parameters in the `Named` derive macro, removing all of the custom implementations (some of which were wrong!).
2) Make `Named` purely about deriving a `Named` implementation (which is what it always should have been!). Type registration is purely for diagnostic purposes, and it is exposed through a separate macro.

This also helped simplify `Named` itself: for example, we no longer need to expose the option to turn off `dump` (what is dump??!), removing some otherwise befuddling behavior.

Derive macros should be purely about deriving an implementation, and nothing else. This fixes this for Named.

Type registration is useful and important, but we should separate concerns. We should think through ways of reducing the burden of the various macro incantations we're creating left and right here. (To be fair, what we're doing is complex, and Rust doesn't provide many other tools to enable this kind of 'dynamic' behavior.)

Reviewed By: shayne-fletcher

Differential Revision: D77910443