actor_mesh: use commactor in cast_slices (#461)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/461

refactor `cast_slices`. the old approach iterated over the input slices and for each rank in each slice, sent a message. now, call `reify_views` on the input slices to produce a single `Selection` (union of ranges). then, cast the message to the selected ranks via the comm actor.

Differential Revision: D77935610