actor_mesh: override cast for SlicedActorMesh (#488)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/488

the sequence of diffs, D77963903, D78030552 and D78104511 have led to the default implementation of the actor mesh trait's `cast` being quite wrong for `SlicedActorMesh`. this is a transitional state; we are working towards a revised design.

this diff papers over the cracks making `cast` on slice actor meshes, "do the right thing".

Reviewed By: mariusae

Differential Revision: D78104511