tensor engine {handle_cast, handle} -> handle (#499)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/499

We unified the message handling logic in rust, so we can keep that unification for the python bindings. We keep the handle_cast variant, rename it handle.

We provide a definition for cast coordinates even when a message wasn't cast: the message is considered to be the  rank of a zero dimension shape.

Reviewed By: suo, mariusae

Differential Revision: D78061501

fbshipit-source-id: f66758cdbcacc103436a23e643769b1df33b6ee5