fix pid in log_writer (#500)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/500

Previously process ID was derived wrong by using the process ID of the current process. That is wrong because the log_writer is spawned on the parent process.
This diff corrects it by attaching the Child process's pid to the log writer so this pid can be used directly in generating the Log Object Name.

Reviewed By: highker

Differential Revision: D78105348

fbshipit-source-id: 42fbbbaee04e03e7c9690d351dbe6450971bef59