add monarch_no_torch wheel and fbpkg (#502)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/502

It's extremely useful to have have a version of monarch that does not require linking against torch.

torch's ABI is not stable. This means that we need to build a wheel for each torch version that we plan to link against and match them up correctly, which is not always convenient when we are trying monarch against all sorts of random envs.

So we add a monarch_no_torch wheel and fbpkg, to have a portable wheel we can install (pretty much) anywhere.
ghstack-source-id: 295691115

Reviewed By: highker, colin2328

Differential Revision: D78168948

fbshipit-source-id: d3f693bc360ce173197655bfcf9d7e2cd568ebbf