cut log spam (#524)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/524

The amount of spew when just trying to use monarch as a user was crazy lol.
Should fix https://github.com/pytorch-labs/monarch/issues/411.

This creates two layers:
- A file log layer
- A stderr log layer

They are configured separately depending on environment. The file one is controlled by RUST_LOG, and the stderr one is controlled by MONARCH_LOG.

For now, both are always used, even if redundant.

You can look at the proposed defaults on an environment basis, but for local/mast we have the same config: errors to stderr, info+ to file.
ghstack-source-id: 295862670
exported-using-ghexport

Reviewed By: dcci

Differential Revision: D78220912

fbshipit-source-id: 2077360f6408fea5413393d895745e1ce90ee218