handling channel not established case for remote allocation (#566)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/566

Remote allocation keeps tracking of proc state by using a hash set. It
adds and remove procs based on the state. It also does sanity check on exit.
However, it misses a case when nothing is allocated. This doesn't leave a clean
exit state. On the contrary, this should be an error.

Reviewed By: pablorfb-meta

Differential Revision: D78475793

fbshipit-source-id: 0b7b472d85e3619f8fec4fa549c8c8d675429351