at least show something when init fails (#586)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/586

The goal is to show why `__init__` fails when creating a actor mesh. This is not a proper fix as the right solution should follow T229200522. I was halfway into the issue until I found T229200522. So pzhan9, I will leave it to you to get the right fix. This diff is a mitigation.

Reviewed By: suo

Differential Revision: D78606675

fbshipit-source-id: 39cfdfa53d75d434c3da66ffcf6a8981225dc1eb