Use register_signal_cleanup to shut down procs in bootstrap (#579)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/579

This fixes `test_allocator_stuck_job`.

Previously we had a problem where if we branched into any of the message handler branches where we called `std::process::exit()` the loop would break and we would not be able to receive any signals. This would result in us being unable to kill the process with SIGTERM which is why the broken test would hang. Registering our callback with the global signal handler will ensure that  even once the loop breaks we can still listen for signals and handle them

Reviewed By: vidhyav

Differential Revision: D78516491

fbshipit-source-id: 5b90bc0e2486d1addcf98e734e227bd19f2b6ee1