tensor engine, send result of indirect call locally (#621)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/621

To be able to have an actor _return_ objects containing tensors to the tensor engine, we need to communicate the result back to the tensor engine without serialization.

This switches the indirect call logic to report the result on a PortHandle rather than as a python message.

To do this, I needed to generatlize `Port` in python to be anything that accepts the `PortProtocol`. We now have three concrete implementations: (1) Port, (2) DroppingPort (for when there is not response port), and (3) LocalPort which is a wrapper around a PortHandle<Result<PyObject, PyObject>>.
ghstack-source-id: 298170006
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D78749865

fbshipit-source-id: 306a3b381e81ef139f4cbaf1196745c928ae7047