Refactor dependent error handling (#619)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/619

Next diff will add rref calls for tensors so I am getting the call_function logic consistent so that a call function that is not calling a function can be implemented. This moves all the handling of error propagation for call_function into `try_define`, similar to python backend instead of handle_results.  It is possible other cases in stream.rs also should be converted over to use try_define as well to ensure consistent error propagation.
ghstack-source-id: 298043755
exported-using-ghexport

Reviewed By: suo

Differential Revision: D78683483

fbshipit-source-id: b33d09776d4f22fa26e13fbac2d4172e25e17874