: selection: enable promotion of any(true()) (#643)
Summary:

the key contribution of this diff is to enable:
```python
Selection.all() and Selection.any()
```
which provide succinct python expressions meaning "route to all nodes" and "route to a random node" in a mesh.

this is supported is by a newly added function `canonicalize_to_dimensions`, which ensures that:
- `*` expands to `*, *, *, ...` (matching the dimensionality of the target slice),
- `?` expands to `?, ?, ?, ...` as needed

secondary changes are refactors:
- canonical embedding of 0d -> 1d slices is now performed in `RoutingFrame::root`
- `promote_terminal` is removed in favor of `canonicalize_to_dimensions`.

these changes shift selection normalization to an up-front pre-processing step, rather than performing adjustments "on the fly" during routing. this simplifies `next_steps` which now has a clearer, single responsibility.

Differential Revision: D78934953