tensor engine, implement rref on actor endpoints (#625)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/625

Allows actor endpoints to take tensors as arguments and return them as monarch.Tensor objects.

Actor endpoints that do this must have a propagator function defined.

"cached" does not currently work because the caching machinery cannot make the actor call.

Introduces `as_endpoint()` which can take something that was not annotated on a class and turn it into an endpoint. This is useful when needing to define a per-actor propagation function for rref calls because there otherwise is nowhere to put it.
ghstack-source-id: 298650755

Reviewed By: suo

Differential Revision: D78782587

fbshipit-source-id: 079fa58143fc32e29edb8147b5f226b810a787f7