Add @endpoint(explicit_response_port=True) (#634)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/634

This option exposes an option that rust actors have always had to get the response port as an argument instead of using the return value of the function as the response.

Having this feature allows synchronous actors to defer responding to a message and process other messages, making them as expressive as asynchronous actors which accomplished this by awaiting.

Enjoy the giant overload copypasta to appease the python typecheckers.
ghstack-source-id: 298823614
exported-using-ghexport

Reviewed By: suo

Differential Revision: D78901486

fbshipit-source-id: ad8371703cfe9c4ed711654a49c39a34324c99f5