Introduce pytokio.rs, run python coroutines on tokio event loop (#662)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/662

The purpose of the diff is to prepare the code for the next diff, which will make ProcMesh, Alloc, and ActorMesh into "remote references" rather than values returned as futures. That is .spawn(), proc_mesh(), etc. will no longer return futures but instead just directly return the handle. This is because having to wait on the creation adds significant challenges for synchronization (do I init this class by first spawning all this stuff and then .get it? How do I write code that works both in sync/async?). Futhermore, awaiting the result of spawning it adds little value. It only says that the object existed some time in the past. However, because of the network and time, they might be gone errored since then anyway.

To create these remote references, we need to hold handles in Python to incomplete work (allocating procs, spawning actors) in the tokio runtime. We already have this functionality in the form of the PythonTask object (tokio coroutine that returns a python object). However, we also need to execute some python code after these things complete to finish initialization (e.g. create the proc after creating the alloc, or do a send after a spawn completes). This PR adds the abililty to create a PythonTask (i.e. a tokio coroutine) that runs a _python async function_ on the tokio event loop. These coroutines do not use `asyncio` or `asyncio.Future`. Think of them as python bindings directly to tokio.  They are implemented by having rust code step the Python coroutine until it hits tokio::Future, and then having the rust async function await it and send it back to Python.

This change replaces the blocking implementations  'emulated' via coroutines in previous PRs. Now all monarch-private work is done in the tokio event loop.

Even before switching to remote references, this switch provides a number of fixes:

1. synchronous uses like .get() previously did not actually start executing until the `.get()` was issued. Now that run concurrently with client code (because they are running on the tokio loop).
2. Initialization of debug, rdma, and setup actors now no longer blocks a user's asyncio loop.

Note: it is important that internal functions do not `await` our Future objects. They should instead only `await` PythonTask objects. The error messages when this happens can be clarified as a hint when it is messed up.
ghstack-source-id: 299200746
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D79058512

fbshipit-source-id: f27871424efee9f6ec951040832a02cd36893d76