Fixed a bug in the channel that was not detecting out of seq messages (#663)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/663

Now, we break the session if we detect an out of seq message. This was disabled because of a failing test. But it turns out the test is wrong. The test stops NetRx but that should kill NetTx because netrx failure results in the entire session state nullified on  the receiver side.

Reviewed By: pzhan9, mariusae

Differential Revision: D79044526

fbshipit-source-id: e83c797d8e410b7b3185b42b59367ed08aff5e43