ndslice: introduce views (#709)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/709

This adds `View` to the ndslice module.

A view is a collection of ranks, projected into an extent. It is represented as a set of labels and a slice. (Subject to extension later.)

Currently, views can be constructed by subsetting an existing view or an extent. Specifically, we support specifying ranges for each dimension.
ghstack-source-id: 299939121

Reviewed By: shayne-fletcher

Differential Revision: D79372445

fbshipit-source-id: 9f6e75427c22b562eb42a5e1acbd1fd1ad454dd0