use serde_bytes for benchmark payload struct (#716)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/716

I am actually not 100% sure whether we want this. Open to discussion.

`serde_bytes` would significantly reduces [the serialization cost](https://fburl.com/code/sd0lv8zc), and make the test measure mainly the transport cost. This results in meaningful throughput improvement: P1888248972

I think using `serde_bytes` on the payload would make it closer to how message structs are serialized in monarch, e.g.:

https://www.internalfb.com/code/fbsource/[92bbcaca2f8cddc859079c1028114a97f615e10f]/fbcode/monarch/hyperactor/src/data.rs?lines=307%2C311-312

which makes the test results closer to the production scenario.

Of course if we want to make a heavy ser/desr workload as part of the test, then I also see the point of not using it.

Reviewed By: pablorfb-meta

Differential Revision: D79383299

fbshipit-source-id: c0e08c7905b17b77ec012209b567e2062a280624