fix broken tests (#740)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/740

Start to see tests in the 2 `test_python_actors` start to fail with `FATAL` exception.

https://www.internalfb.com/intern/test/562950169060652?ref_report_id=0

This is an indication there is a race at the shutdown, and the race leads to acquiring a released GIL.

I guess one diff landed last Friday make the race more frequently. This error can be prevented by stopping the proc mesh at the end of the test. Ideally we should stop the proc mesh anyway to release resource.

Reviewed By: dcci

Differential Revision: D79526366

fbshipit-source-id: da9f1ec81f5e3c065bf883b6aa3a67f16e69c2aa