Polymorphic Future await? (#757)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/757

Previously if on the tokio event loop we forced our internal uses to directly await the PythonTask.  This leads to a few places where we end up with a Future but want to await it from a coro also on the tokio event loop.

We have the option of making `__await__` work in either case: on a asyncio loop, create a python future, on a tokio loop spawn/await.

If we want consumers to create Future objects themselves we probably want this. If we want tokio event loop things to be an implementation detail of monarch, we probably do not want this.
ghstack-source-id: 300750788

Reviewed By: allenwang28

Differential Revision: D79596925

fbshipit-source-id: 6ccde7e9762fce85b83e169a6157be1164e7ecff