root_client_mailbox: open ports without a proc (#683)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/683

To get ready to make proc_mesh/actor_mesh remote references, we need to be able to open ports on the root client without having to wait for the proc to exist.

This diff creates a global Mailbox hooked up to the global router that can be used on the root client to open ports.

On actors the mailbox is looked up via the monarch context set while the actor is running. On the root proc, we use the new global mailbox.

This makes it possible to create a function to create a port without other arguments. This function is actually a class named Channel[T] because of python typing reasons: there is way to pass a type argument to a function in Python. Called channel because that is the rust equivalent. Channel[T].open() returns what used to be called PortTuple, but is now just a real tuple.
ghstack-source-id: 300749961
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D79215096

fbshipit-source-id: 5ace368b0813467969cc0f43ef587ab4775cbf50