Pass cast rank to python actor (#747)
Summary:
Pull Request resolved: https://github.com/pytorch-labs/monarch/pull/747

When casting to a sliced mesh, the actors rank on the sliced mesh is different from the rank on the root mesh.

Currently, the root mesh's rank is passed to Python actor. [That is wrong](https://www.internalfb.com/diff/D78355743?dst_version_fbid=1460199675405905&transaction_fbid=1279216557057466). We need to pass the rank on the cast mesh. If the cast mesh is a sliced mesh, then it should be sliced mesh rank.

This diff fixes this.

Reviewed By: mariusae

Differential Revision: D79530146

fbshipit-source-id: 0721964e3c532b82d91e48eb189ab97fc7d5435a