channel: allow ":" as a channel-type delimiter (#798)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/798

We currently use "!" to delimit between channel type and its address. The main issue with "!" is that it is not shell-friendly, as Bourne shells will expand "!" into a history entry.

With this change, we allow ":" to be used instead. With the previous change (requiring channel types to be specified), we can parse these two cases unambiguously.

The ChannelAddr `Display` implementation is adjusted to emit ":"-delimited channel addresses, too. It is the new preferred syntax.
ghstack-source-id: 301740066

Reviewed By: shayne-fletcher

Differential Revision: D79846852

fbshipit-source-id: 1304564d9e99b229172cd63dc97aa273b9424a01