simplify supervision propagation; unhandled events always cause failure (#784)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/784

Currently (local) supervision can propagate events without also killing an intermediate actor.

This is 1) wrong; and 2) complicated.

Instead, we treat an unhandled supervision event as an actor failure, and then reduce the propagation paths to one: that of an actor failing.

In order to retain accurate attribution, we add a "caused_by" field to the actor supervision events.
ghstack-source-id: 301725856
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D79702385

fbshipit-source-id: ccce9e04b206973b6a9b372180a967bbffda79a5