(monarch/tools) Add force_restart argument to get_or_create command (#783)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/783

`force_restart=True` will kill the existing job and force it to create again. Useful when we want to redeploy the server to reflect any changes in the local conda/workspace that are not eligible for CodeSync.

NOTE: this is tangential to setting `workspace=None` which disables local conda AND workspace building into an ephemeral image.

Reviewed By: ahmadsharif1

Differential Revision: D79754696

fbshipit-source-id: 085aa2f1633d5b4ebac375a8494fc754a532d4a1