De-future spawning actors (#813)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/813

This commit switches actor spawning to no longer return a Future. ActorMesh becomes a DeprecatedNotAFuture to keep existing code working.

The implementation delays the send part of endpoints until the actor has spawned. I expect this to have some performance impact, which needs to be measured before we land this.

However, this implementation is not what we expect to actually use: the ability to spawn actors/return a ref immediately will be made part of the rust runtime where the overhead will be substantially less.
ghstack-source-id: 302498028
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D79938417

fbshipit-source-id: f2606bc82087d6aff4f31f826cbad80615cc8982