Move casting protocol into rust (#851)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/851

This moves the logic for deferring sends into Rust to avoid slowing down actor sends. Increases send throughput ~30% from the python approach. It is just a little faster than what came before the python approach because this also eliminates a layer of python wrappers for dispatching sends. I suspect there is tons more to to be gained removing more layers of python between the user calling an endpoint and issuing a send.
ghstack-source-id: 303184134

Reviewed By: suo

Differential Revision: D80037834

fbshipit-source-id: 78c2b59f5c919806b76a5fd56e435fb1a93b4719