Allow proc_mesh.stop() after actor_mesh.stop() (#877)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/877

Currently calling proc_mesh.stop() will fail if an actor_mesh spawned on this proc_mesh has been stopped.

This is because the python-binding wrapper for ProcMesh (TrackedProcMesh) holds onto a map of SharedCells containing each RootActorMesh spawned on it. When an PythonActorMesh is stopped, the RootActorMesh is dropped, but does not update the state of the TrackedProcMesh to remove this. When a ProcMesh is stopped, it will attempt to discard every SharedCell containing it's RootActorMeshes, but will return an error if any RootActorMesh has been dropped (The error we are seeing)

We can make it such that stopping the PythonActorMesh will update the state of the TrackedProcMesh, but this would involve plumbing through references of TrackedProcMesh and related state to PythonActorMesh

A simpler solution, it to simply allow the TrackedProcMesh to discard the errors related to attempting to discard a SharedCell where the RootActorMesh has already been dropped

Reviewed By: ahmadsharif1

Differential Revision: D80269385

fbshipit-source-id: a2b92f8fdba99e5a0cbec31dc8dfd351c4be2477