Pulling changes of grpo_actor from docs version to original copy (which has edit history). (#921)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/921

This diff pulls in changes from the doc-team's copy of grpo_actor.py to the original, and removes the copy.  This is to prepare for the next commit, which will move the original (w/ changes) to the doc examples area so we have the same thing but with the change history recovered.

Reviewed By: shayne-fletcher

Differential Revision: D80534401

fbshipit-source-id: 4c0970f9e1abfeda0765e9ce34e38a57ffc139cf