channel-level ping-pong benchmarks (#906)
Summary:

This is an attempt to do an apples-to-apples comparison to P1903314366, to eliminate any non-channel related overheads.

The results replicate previous findings: that our throughput is hampered by excess data copies (either outright or through growing buffers in the encoding stack), and of tokio-level network i/o overheads. Both are being addressed. These benchmarks should help to serve as validation as this work lands.
ghstack-source-id: 304032339
exported-using-ghexport

Reviewed By: highker, vidhyav

Differential Revision: D80260732