net: zero copy framer (#907)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/907

This change introduces a zero-copy framer, which will also be easily extendable to doing vectorized framing of multipart messages.

We eschew tokio's framer (which relies on queuing for cancellation safety), in favor of a simple(r) implementation: the reader maintains a simple state machine, while the writer requires the caller to maintain an explicit write state (since this has to be driven across selects in order to be made cancellation safe).

In this way, we have an easily hackable framer that does not introduce additional queues.
ghstack-source-id: 304032337
exported-using-ghexport

Reviewed By: vidhyav

Differential Revision: D80365228

fbshipit-source-id: c46c10ddbcc32154e5ff9461d773611a3c97945e