Make Encoded::Bincode and Encoded::Json cheaper to clone (#941)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/941

We probably will use `Encoded::Multipart` more extensively in the future with D80704705. But it does not hurt to improve `Encoded::Bincode` and `Encoded::Json`.

Currenlty, clone `serde_bytes::ByteBuf` would result in a copy of the underlying `Vec<u8>`. `bytes::Bytes` on the other hand would be zero-copy.

This change should improve where `Encoded` is cloned. But specifically I write this diff for casting, where a large number of clone could happen in a loop:

https://www.internalfb.com/code/fbsource/[ab89a2b4ce2264912c5b0dafab3a97829b0b834d]/fbcode/monarch/hyperactor_mesh/src/comm.rs?lines=279-281%2C290

Reviewed By: shayne-fletcher, vidhyav

Differential Revision: D80657648

fbshipit-source-id: e9a954c7de68dc29f7622be0e33bf68ba20d5d7c