Make MessageEnvelope takes a vec of errors (#971)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/971

Right now, `MessageEnvelope` only allows up to 1 undelivered error. But very often the message could be associated with a series of errors. For example:

1. The 1st error is due to the message cannot be delivered to its dest;
2. The 2nd error is due to the message cannot be returned to its sender.

We want to log all of them in order to get the full picture of why it was not delivered.

This diff makes this change. It also improves some call sites with this new capability.

Reviewed By: shayne-fletcher

Differential Revision: D80846618

fbshipit-source-id: a914adcfb1b4732cf9448fc835c6ff6cf0f8f397