Improve logs in net.rs (#1004)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1004

The main change is to standardize how to identify a log statement. Specifically, it uses `log_id` as the prefix of logs if applicable.

`log_id` in `NetRx` is

```
let log_id = format!("session {}.{}<-{}", self.dest, session_id, self.source);
```

and in NetTx is:

```
let log_id = format!("session {}.{}", link.dest(), session_id);
```

Ideally we want contextual logging to automatically append this prefix so we do not need to manually annotate every logs. But we cannot wait for that since we need the standardized id now in debugging to correlate logs.

Reviewed By: shayne-fletcher

Differential Revision: D81132941

fbshipit-source-id: 82dcefb0cd21ac59bd0d59921dcdd2cac826828f