channel example: make it into a flexible stress test (#1002)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1002

With this change, we can stress channels better by running the server and client independently (and potentially on separate machines). By default, it runs in "infinite" mode, reporting stats every 1000 messages.

The idea is that we can use this as a test bench to try to replicate issues we see in production with channels.
ghstack-source-id: 305992192
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D81099840

fbshipit-source-id: c6e3166c4fce1c2c9afdf7fc80abbc551047be5e