Don't wait 1/2 second every time we launch a new proc (#1104)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1104

tx.send waits for a response that the send was received. That response seems to only be sent at MESSAGE_ACK_TIME_INTERVAL (.5 seconds). This means that pretty much every launch of a process has a sleep(.5) seconds in it.

This changes it to simulatenously monitor for a send error or the response, eliminating the sleep.

Gaze at the hideous tokio::select invocation needed to do the right thing here and question whether the efficiency of rust is never achieved because doing the right thing is just too hard.
ghstack-source-id: 307628873

Reviewed By: mariusae

Differential Revision: D81709502

fbshipit-source-id: e65918bd94499e7e7ee4b7e28b60503a6c527df3