mesh: refactor and clean up allocation
Summary:
This is in preparation for host meshes. We refactor `ProcMesh::allocate_boxed_inner` into several methods, and provide a new `Router` type to manage the global router, and provide a sensible API for it.

This will then allow us to reuse this code in the HostMesh context more easily.

Differential Revision: D81552705