mesh: introduce ValueMesh, ProcMeshRef::status (#1087)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1087

ValueMeshes let us represent a mesh of values. We subsequently use this to represent the status of a proc mesh.

This ought to be made more generic. Ideally it would behave like an ordinary Rust collection (with the attendant implementations of FromIterator, etc.). The challenge presented is that not all iterators are valid ValueMeshes, since we require point compatibility and completeness. We also can't define a FromIterator implementation for Result<ValueMesh<T>, Error> because of the foreign type restriction.

Another optimization we'd like here is to have a run-length encoded representation (for types that implement `Eq`).
ghstack-source-id: 308284568
exported-using-ghexport

Reviewed By: shayne-fletcher, highker

Differential Revision: D81612446

fbshipit-source-id: d51c902a91a4aafa768db6628c4f4e1478089bb6