mesh: core v1 API: HostMeshRef and ProcMeshRef (#1079)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1079

This puts in place the core "v1" API, consisting first of HostMeshRef and ProcMeshRef. These are now cloneable entities that implement `ndslice::View`, so that they may be subsliced, partitioned, etc. in a consistent and principled way. They are defined as logically a set of (compactly) ranked Host and Proc refs.

This does not yet include any functionality; rather this diff sets up the basic structure of these APIs.

We also introduce `ndslice::view::Ranked` as an implementation trait that implements `View` for ranked collections (as are common).

Finally, this is all put into a new module hyperactor_mesh::v1, which is meant to contain (and distinguish) the "v1" API for hyperactor_mesh. Once we are migrated over, we can move these modules into the base hyperactor_mesh module.
ghstack-source-id: 308284565

Reviewed By: shayne-fletcher

Differential Revision: D81590896

fbshipit-source-id: 2cf71515cd380974e1805ac8ed06d74ba3f8f3c8