mesh: refactor and clean up allocation (#1075)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1075

This is in preparation for host meshes. We refactor `ProcMesh::allocate_boxed_inner` into several methods, and provide a new `Router` type to manage the global router, and provide a sensible API for it.

This will then allow us to reuse this code in the HostMesh context more easily.
ghstack-source-id: 308284560

Reviewed By: shayne-fletcher

Differential Revision: D81552705

fbshipit-source-id: 190bf0a947fa09c3fcf0598e50ab386caa78c386