Back out "Don't pre-import torch" (#1137)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1137

https://github.com/meta-pytorch/monarch/pull/1112 broke a workflow used internally,
where the monarch python library is prepended to PYTHONPATH instead of installing
the wheel.
As such, importing a library like torch is necessary before importing anything which needs
the C++ bindings like `_rust_bindings`.
For now, revert to avoid breakages. We'll find a better fix soon so we don't
need this import.

Reviewed By: zdevito

Differential Revision: D81980772

fbshipit-source-id: 476e90fb13abdb1b42c745e3f9a7255fd0f50afc