Improve NetTx exiting log (#1060)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1060

Currently, if there are too many seqs in `Outbox` or `Unacked`:

https://www.internalfb.com/code/fbsource/[315bec9cb3e7f40c08421ccefd1dfd489c3c55d3]/fbcode/monarch/hyperactor/src/channel/net.rs?lines=225%2C230%2C319-320

Their `Debug` string will be too long.

This diff wraps `MessageDeque` with a new type, and provide a custom Debug impl for that new type, in order to improve the debug string output.

In addition, this diff also adds a `sent_at` field to `struct QueuedMessage` so we can log when the message was written to the stream in the log.

A example log with this change will look like: P1937297583

Reviewed By: mariusae

Differential Revision: D81486270

fbshipit-source-id: defa2c0dacd7abfea149206a3ba34d636571d727