Use SmolString optimization to reduce heap allocations on logging path (#953)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/953

Don't directly create scuba logging objects on the hotpath, instead append to a vector of key-value pairs and do the logging on the background thread.

In testing, this lowers the variance of logging on enter/exit of spans dramatically (as much as 20us in some places).

Since most of our strings are < 23 chars long, using SmolString also reduces the number of heap allocations required. In the optimal case, only a single heap allocation is made (the Vec::with_capacity) for each logging call.

Reviewed By: mariusae

Differential Revision: D80573843

fbshipit-source-id: fed54dc2b0749698d5a0e3db84f538ba5088abd9