switch over (mostly) to the new context module (#1183)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1183

This was a little more involved than I thought it would be. Primarily, this was driven by porting the macros to use `&impl context::Actor` instead of capabilities. We then converted everything downstream of this. The one exception is the RDMA manager, which is Rust code, invoking macro-genned code, dispatched through Python. I left a `_deprecated` version of each client method to use the old caps (which can be provided by `Mailbox`). Once samlurye's work plumbing through Rust Context/Instance is done, we can switch these over as well, and remove the legacy codegen.

Some other casualties include some mock-oriented tests in multiprocess (but we're deprecating this crate anyway).
ghstack-source-id: 309592415

Reviewed By: shayne-fletcher

Differential Revision: D82248674

fbshipit-source-id: dea6e9398822f90f362892067e35c37e48e2c5e1