Fix race in proc shutdown (#1203)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1203

This diff fixes a (highly unlikely, but still real) race during `Proc::destroy_and_wait` where we might call `unwrap()` on an `InstanceCell`'s actor task handle before the handle is actually set, causing a panic. It replaces `unwrap()` with `wait()` so that we are guaranteed to have an actor task handle. See the comment in code for why this should be safe to do (and not hang forever).
ghstack-source-id: 309646341
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D82355158

fbshipit-source-id: 7ac170920dd58a0fef789535057ce117dfe65ae8