ProcessProcManager (#1201)
Summary:

This implements a ProcManager for external processes, following a simple protocol.

Each process is launched with the following environment variables:
- HYPERACTOR_HOST_BACKEND_ADDR: the backend address to which all messages are forwarded,
- HYPERACTOR_HOST_PROC_ID: the proc id to assign the launched proc, and
- HYPERACTOR_HOST_CALLBACK_ADDR: the channel address with which to return the proc's address

The launched proc should also spawn an actor to manage it - the details of this are
implementation dependent, and outside the scope of the process manager.

The function [`ProcessProcManager::boot_proc`] provides a convenient implementation of the
protocol.
ghstack-source-id: 309631190
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D82342107