Reimplement python bindings for `Instance` and `Context` in terms of the actual underlying instance (#1222)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1222

Now that `Instance<A>` has only cloneable state, we can implement the python `Instance` and `Context` objects in terms of `Instance<A>` (rather than just the mailbox and actor id). This is necessary because of the new `context::Actor` trait that replaces the caps system -- we need access to the actual `Instance<A>` in python so that we can spawn new proc meshes/actor meshes. Unfortunately pyo3 doesn't support generic types in structs, so the `PyInstance` struct actually contains a new struct called `ContextInstance`, which multiplexes `Instance<PythonActor>` and `Instance<()>` (for the root client context).
ghstack-source-id: 310134496
exported-using-ghexport

Reviewed By: zdevito, mariusae

Differential Revision: D82476081

fbshipit-source-id: a2aa7edff99635a70ee319fb0bdf330fb3dc018f