mesh: host agent, v1 bootstrap, resource management behaviors (#1228)
Summary:

This change introduces the HostMeshAgent, which is an actor that manages a `Host`, creating and observing procs through the `Host` API.

A HostMesh is a ranked collection of HostMeshAgent references.

We extent [`hyperactor_mesh::bootstrap`] to be `Host` compatible, through a custom `ProcManager`, which serializes its configuration into a single environment variable, providing better type safety and simplicity.

This change also introduces a "behavior" (i.e., a set of messages) for managing resources. The HostMeshAgent implements these messages for managing procs. We will later extend the ProcMeshAgent with the same messages, providing a uniform API for managing the lifecyle of the resources across all meshes.
ghstack-source-id: 310344933
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D82534107