Fix another race condition in proc shutdown (#1232)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1232

This fixes the tests in T238124137. The problem is that if `proc.destroy_and_wait` is called from inside an actor, then after aborting that task's handle, `destroy_and_wait` never makes it past this await point: https://fburl.com/code/xzgtv3mj. This wasn't a problem before because the await point was added recently. The solution is to delay aborting the actor task handle that `proc.destroy_and_wait` was called from.
ghstack-source-id: 310876263
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D82558878

fbshipit-source-id: 05130641bd9beb9e7203876485270391d912f914