bootstrap: address some snagging issues (#1319)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1319

this diff removes the `Agent` associated type from `ProcManager` and replaces it with a `ManagerAgent<M>` type alias via `ProcHandle::Agent`. `Host::spawn` now returns an `ActorRef<ManagerAgent<M>> `and the redundant `type Agent = …` lines in the local and bootstrap managers are dropped. a `Display` implementation is added for `ProcStatus` that prints pid and uptime using humantime, and corresponding display tests are introduced. documentation and error text are updated to refer to `Ready` rather than `Running`. humantime is added as a dependency in BUCK and Cargo. cleanups include dropping an unused variable and trimming some test imports.

Reviewed By: mariusae

Differential Revision: D83110957

fbshipit-source-id: 2eaa1f672ad7618ee73aa12635a44ec06c7b66da