jobs api: adding trait and local (#1301)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1301

Beginning the implementation of RFC: https://docs.google.com/document/d/1QJ1WeRYjDElJ_bySgekWVMujdCXc3oIJRB0Cvkjj-UI/edit?tab=t.0

This diff adds:

* The Jobs Trait
* LocalJob

The LocalJob can schedule as if it were a batch job with .apply taking a client_script. This is useful for testing and for consistency with remote batch scheduling but probably won't get much use.

Next up is to pull in the code from monarch.tools so that we can use the Jobs api with the torchx launchers we already support.

Reviewed By: mariusae

Differential Revision: D83015058

fbshipit-source-id: 8c6136f7144008a84adfe03c7dfcb7a2977921c1