Should not restrict RdmaController on RDMA instance (#1164)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1164

Currently, the system will fail if the RdmaController runs on an instance that does not support RDMA. For example, if you launch a MAST job with RDMA-enabled instances, your local environment is still required to support RDMA, because that is where the RdmaController runs by default. (Sidenote: It is not trivial to change the RdmaController to run on the MAST mesh either, because it is a singleton, whereas MAST mesh jobs run as multiple actors, and slicing a mesh is not supported yet.)
However, there is no reason for the RdmaController to require running on instances that support RDMA. This diff contains the changes to remove that restriction and add the check to `RdmaManagerActor::new` instead.
Even aside from the perspective of restrictions, it seems like a bug to me that the RdmaManagerActor is using the configs (which is fine today because it’s just the default) from the controller’s mesh instead of its own mesh.

Reviewed By: dstaay-fb

Differential Revision: D82124576

fbshipit-source-id: 90a22dd7c0aa53a099024f89c247c832acba79d8