Add casting support to resource::GetState messages (#1262)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1262

Part of: https://github.com/meta-pytorch/monarch/issues/1209

Use casting to implement the `supervision_events` API instead of iterating over all
ProcMeshAgents. This will scale better as the size of the ProcMesh increases.
This requires making some trait bound changes to `resource::GetState` so that it can be casted with.

Unfortunately, the Actor name for the mesh agent is not compatible with the v1::Name struct
due to the missing uuid. Make `v1::Name` an enum to allow reserved names to be used for
things like ActorMeshes.

Also, minor improvement: make ActorMeshRef::supervision_events not take a Name, we can assume
it is for the current mesh's name.

Reviewed By: mariusae

Differential Revision: D82687236

fbshipit-source-id: deac40638aa238feeb265ccd2cc949c65b2db3f6