Rename supervision_events to actor_states (#1332)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1332

Part of: https://github.com/meta-pytorch/monarch/issues/1209

Rename from supervision_events to actor_states, and return the whole State object.
This is more in line with what we want, which is an API that can accumulate statuses across the whole
mesh to continuously monitor when a member transitions to unhealthy.

Reviewed By: shayne-fletcher, mariusae

Differential Revision: D82761485

fbshipit-source-id: 7db9bc1a9601e82f2996407c8d46bd895bf74922