Update Handler macros to use `context::Actor` (#1349)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1349

Motivation: In `Rust V1`, we will make the sequence assignment logic an actor-level logic. Specifically, the sender actor is responsible for assigning its message's seq numbers. The implementation plan is to encapsulate that logic inside sender's `context::Actor` object (see D82983004). As a result, we need to surface `context::Actor` to all the send call sites, so the sender's sequencing capability is accessible. These callsites include `OncePort/PortRef.send`, and `PortHandle.send`, `ActorHandle.send`, `ActorRef.send`, and the corresponding callsites on the python side.

For this diff, it specifically updates the macros.

Reviewed By: mariusae

Differential Revision: D83371710

fbshipit-source-id: 994c776c93b449e57e04ef83c38ce9badaefc5f3