MASTJob implementation (#1351)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1351

Adds a implementation of the JobTrait for mast based on the already existing get_or_create logic in monarch.tools.

Works for local/notebook uses. Does not yet support launching with a training script because I have to figure out how to do that with the mast APIs.
ghstack-source-id: 312926533
exported-using-ghexport

Reviewed By: colin2328

Differential Revision: D83377414

fbshipit-source-id: 42d2298500fbaebf0bac4125e8dc8595fd603446