tighten the usage of bind_to (#1437)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1437

`PortHandle.bind_to` is not meant to be a public method. This diff makes it `pub(crate)`, rename it to `bind_to_actor_port` and remove its `port_index` parameter.

https://www.internalfb.com/code/fbsource/[5e07de58da822924da9c2108a46ba4cd173a75c8]/fbcode/monarch/hyperactor/src/mailbox.rs?lines=1600%2C1613

However, this method is already used outside the `hyperactor` crate in a couple of places. This diff hides those usage behind a new method `Instance::<()>::bind_actor_port`.

This diff also removes `Ports::bind_to` and fold its logic to where it is currently used.

https://www.internalfb.com/code/fbsource/[5e07de58da822924da9c2108a46ba4cd173a75c8]/fbcode/monarch/hyperactor/src/proc.rs?lines=1862%2C1958

Reviewed By: mariusae

Differential Revision: D83931993

fbshipit-source-id: 35dce72c235ec54b883df4c12c44d6141446db2c