Update the `bound` field within `PortHandle.bind_default` method (#1439)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1439

Currently, `PortHandle.bind_default` does not save `port_index` in the handle's `bound` field after it binds the port. This diff changes that. In this way, we can know whether the handle object is bound, and bound to which port, by using `PortHandle::location` method.

This change is a behavior change to another method, `PortHandle::bind`. Previously, `bind` can still mutate the `bound` field aster `bind_default` was called. After this diff, it cannot. The new behavior is documented in the new test,  `test_bind_port_handle_to_default`.

Reviewed By: mariusae

Differential Revision: D83934172

fbshipit-source-id: e761c1f697dc26b219a01f45046a1421ca0bd065