Fix up failing tests by adding Tcp::Localhost mode (#1479)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1479

This diff adds support for tls localhost to fix T232884876. I go into detail in https://fb.workplace.com/groups/1434680487636162/posts/1441470780290466/, but basically, the hostname can't be bound to for certain environments. This generally takes the approach discussed with mariusae, though I ended up doing localhost instead of unspecified. This is because the tests were already set up with localhost; it was just being ignored.

This is a bit of a short term fix, as I hope to follow up by allowing the hostname to get sent to workers while using unspecified locally. DNS is "more correctly" resolved on the worker machines, due to proxying and load balancing questions, but that's a bigger lift.

Reviewed By: mariusae

Differential Revision: D84094699

fbshipit-source-id: f7ed39a39e6ca153dc0a6e8cbdeb5dbf852e54f8