Add stop for v1::ProcMesh (#1560)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1560

Add in support for ProcMesh::stop for v1 proc meshes.
This sends a Destroy message to the HostMeshAgent that owns this proc mesh.

This only works for owned ProcMeshes, they cannot be stopped from a reference.
The Name for HostMesh procs contained an underscore and required a change to Name parsing,
that only the last underscore separates the uuid

Reviewed By: mariusae

Differential Revision: D84556616

fbshipit-source-id: 9a0393691a601d1e8f9f96886a4f8b1eef21b2b7