Fix assert with undeliverable message from comm actor (#1606)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1606

Fix a panic in PythonActor::handle_undeliverable_message when the "sender" is the comm
actor. We need to update the sender back to the original "self" actor by using the headers
set in the envelope.

After this fix, instead of a panic we get a supervision error:
```
SupervisionError(
Actor ...wrapper_1xYczTZiTdb1[0] exited because of the following reason:
<PyActorSupervisionEvent: ...wrapper_1xYczTZiTdb1[0]: failed: serving
...wrapper_1xYczTZiTdb1[0]: processing error: a message from ...wrapper_1xYczTZiTdb1[0] to
...fail_1am38hE5fnus[0] was undeliverable and returned: Some("send error: channel closed;
multicast error: comm actor comm_1JgvjFbdpnUf[0] failed to deliver the cast message to the
dest actor; return to its original sender's port ...wrapper_1xYczTZiTdb1[0]
)
```

Not very terse, but better than a panic! This also allows any custom override of
handle_undeliverable_message to work.

Reviewed By: pablorfb-meta

Differential Revision: D84952942

fbshipit-source-id: b2cb36600ca89a03e7a85cfccd46ce6bcf2487cf