(4/n)Single FileMonitor per ProcManager (#1570)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1570

Previously we were creating one log file per spawned proc, which led to 2*num_procs of files, this will not be sustainable at Mast scale.

Instead, here we propose to centralize the file appending logic into a single file per host that multilpexes output for all procs under the the same BootstrapManger, which is tied to a single host.

Architecture is
```
Tee ---> stdout
 |    |---> client log stream
 |------------------------>  host log file
```

Notes:
- Currently the log file will be named under a UUID, will follow up to add proc_mesh id.
- User can additionally specify whether they want to prepend the lines from a specific proc with their corresponding rank

TODO:
- Support configurable per rank files
- Name files based on the associated proc mesh

Reviewed By: vidhyav

Differential Revision: D84812385

fbshipit-source-id: 655fafac4eef397129bb49ecced4bfd73ed075c1