Add reason to channel::serve (#1674)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1674

We need to have a way to associate the channel address to why it was served. This diff adds a reason parameter to `channel::serve` and log it along with the address.

We probably do not need this diff in the future if we include "reason" or something similar in the channel's name directly. But before that, this diff is required otherwise we cannot debug channel errors in many cases.

Reviewed By: shayne-fletcher

Differential Revision: D85529357

fbshipit-source-id: d75146766ff93834d5781ee67833084ca5932854