Change ActorStatus::Stopped to not call unhandled (#1675)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1675

Before, calling "proc_mesh.stop()" or "actor_mesh.stop()" would generate mesh
failures that could cause the `unhandled_fault_hook` to be called, even if those meshes
were not used afterwards.

This created a negative user experience, and wasn't making a useful error message. To fix
this, distinguish between requested stops and unrequested failures. The former does not
cause a client crash, and the latter does.
Requested stops will still make proc meshes and actor meshes unusable (cannot spawn new
actors, cannot message existing actors) and the error messages will include that the mesh was
stopped.

Also, on the side, include the mesh name with the mesh failure. This wasn't always included
with the ActorSupervisionEvent (particularly for stopping), and it's very helpful to have when an error occurs.

Reviewed By: mariusae

Differential Revision: D85596621

fbshipit-source-id: 3359c4b1ccb31fb7431a2e5fab686323fd9c84d2