Hook up pytest results for publishing on each Github Action (#1648)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1648

The pytorch test-infra jobs already support submitting test reports for display on a job.
This can improve the user experience when there's a problem with the CI.

Right now this is just for python tests, but cargo also supports the JUnit xml format, so we
can add that later too.

Reviewed By: colin2328

Differential Revision: D85294258

fbshipit-source-id: 0c0fb3b03d25cf00301fb44269b96bd3a6824543