Migrator monarch/python/monarch/_src/actor to pyre-strict (#1710)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1710

pyre had a couple of gaps for how we handled endpoints that interfered with full turning typing on that are mostly addressed with pyrefly.

This is part 1 of a series of diffs to turn on pyre-strict, leaning someone generously on Any's for now, which then will be followed up by enabling pyrefly. Unlike last time, I think this should also generally preserve the existing behavior on how endpoints resolve, but happy to modify that, since how they pass through args can be funky.

Reviewed By: shayne-fletcher

Differential Revision: D85356082

fbshipit-source-id: fcd2f5d2c49cf1166080e97de19a83999ac927be