Back out ": python/monarch: pytorch InputBuffer::add() signature changed" (#1741)
Summary:

Original commit changeset: 43599cd13db9

Original Phabricator Diff: D86025607

this fix appeared to work on the day but didn't "stick". that being so, it doesn't seem that keeping it is worthwhile. better we reset to how we intend the build to work (use torch nightly everywhere) and take it from there.

Differential Revision: D86051511