Back out ": python/monarch: pytorch InputBuffer::add() signature changed" (#1741)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1741

Original commit changeset: 43599cd13db9

Original Phabricator Diff: D86025607

this fix appeared to work on the day but didn't "stick". that being so, it doesn't seem that keeping it is worthwhile. better we reset to how we intend the build to work (use torch nightly everywhere) and take it from there.

Reviewed By: johnwhumphreys

Differential Revision: D86051511

fbshipit-source-id: ecd725d98013e4686f612eea6bfff665db28436a