move from scripts/shaynefletcher (#1757)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1757

moving the hyperactor mesh book into fbcode/monarch/docs/source/books. at this point there's enough structured material (host/agent walkthrough, bootstrapping) that it should live with the rest of monarch docs instead of in my personal scratch space.

this makes it discoverable for people who are trying to understand v1 and we can start iterating on it in-tree with normal review procedures rather than out-of-band.

Reviewed By: mariusae

Differential Revision: D86309851

fbshipit-source-id: 33a475c3ad00810d72fb18400bf89bda651698ce