Version guard for AccumulateGrad::add (#1768)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1768

When running `fbpkg build //monarch:monarch_dev` on 3279935a10 we encounter a compile time error that seems to stem from compiling with different versions of PyTorch. We might want to update our build chain. It seems monarch is once compiled against a third part 2.8.0 pytorch and once against the fbcode version. This version guard helps with this version drift if that's intended.

Reviewed By: pzhan9, zdevito, soulitzer

Differential Revision: D86363543

fbshipit-source-id: cb344e3363595cfe6aed22a27ac8eb8075fae351