ActorStatus::Failed(String) -> Failed(ActorErrorKind) (#1815)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1815

Refactor to get a more structured error to ActorStatus. This deletes the unused structure in ActorErrorKind to make it cloneable/serializable so that we can put it in ActorStatus.

Now that we can represent an ActorStatus that is an SupervisionEvent of a child, we can start to chain/collapse the supervision chain as it bubbles up.
ghstack-source-id: 322758617
exported-using-ghexport

Reviewed By: mariusae

Differential Revision: D86723978

fbshipit-source-id: 3587161f752536ff913cb20ef616bd7c00ae9396