refactor channel::net client loop (#1823)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1823

First, we break down the client loop into several functions, breaking up its monolithic nature.

Second, instead of relying on custom rendering of loop states, we formalize this into a span, which efficiently pulls out the (structured) fields as required. This means we don't pay a formatting/string construction overhead on each iteration, and also make available the underlying fields in our structured logs.

This is just the first salvo in improving how the `net` module is factored.
ghstack-source-id: 322824631
exported-using-ghexport

Reviewed By: shayne-fletcher

Differential Revision: D86783166

fbshipit-source-id: 95e5bf68981f96033774f042153dc6a0af4a5c0b