Add cleanup function to Actor trait and use it from PythonActor (#1836)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1836

Fixes https://github.com/meta-pytorch/monarch/issues/1849

RFC: Add a "cleanup" for actors to run on stop.

This method is invoked after all child actors are stopped, but before the current actor
exits.
If an exception is thrown, it will become an error at shutdown which will propagate.
The cleanup method in python can be sync or async, and must match the syncness of
endpoints. It takes one argument, which is an `Optional[Exception]`. If this was an abnormal exit,
that exception is not None, specifically to the result of `ActorError::to_string` (wrapped
in an exception object).
Later on we may be able to preserve the original exception, if there is one.

The motivation is from actors that want to call `dist.destroy_process_group()`,
as that is one of the most frequent cleanup action in users of monarch. Actors should
*not* call `stop()` on any actor or proc meshes they own. This will be handled automatically
in the future, and they will have already been stopped by the time cleanup
is invoked.

This cleanup is per-actor, not per-proc. So if the cleanup is destroying process-wide resources
(as does "destroy_process_group"), then the actor shouldn't be colocated with any other
actors on the same proc using the same resource.
If the cleanup takes too long, ignore the result and continue with stopping.

ProcMesh::stop() already does a graceful stop of all actors, so this cleanup will be run
automatically when proc meshes are stopped.

Reviewed By: mariusae

Differential Revision: D85624518

fbshipit-source-id: 172eeebf18eddc1a7f5928dcc31efd4cd9120287