mesh: define `resource::Resource`, and mesh::Mesh in terms of it (#1841)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1841

We formalize a resource controller behavior:

```

/// A trait that bundles a set of types that together define a resource.
pub trait Resource {
    /// The spec specification for this resource.
    type Spec: Named + Serialize + for<'de> Deserialize<'de> + Send + Sync + std::fmt::Debug;

    /// The state for this resource.
    type State: Named + Serialize + for<'de> Deserialize<'de> + Send + Sync + std::fmt::Debug;
}

// A behavior defining the interface for a mesh controller.
hyperactor::behavior!(
    Controller<R: Resource>,
    CreateOrUpdate<R::Spec>,
    GetState<R::State>,
    Stop,
);
```

Anything that is a controller should behave-as a `resource::Controller`.

We then formalize the mesh controller behavior as a specialization of a resource controller, by implementing `Resource` for any mesh:

```
impl<M: Mesh> Resource for M {
    type Spec = Spec<M::Spec>;
    type State = State<M::State>;
}
```

This resolves to the same set of bindings and aliases (except we use `resource::Controller<Mesh>` rather than `mesh::Controller<Mesh>`) -- the existing behaviors continue to assert.
ghstack-source-id: 323322438

Reviewed By: shayne-fletcher, vidhyav

Differential Revision: D86905562

fbshipit-source-id: fff583111bcad06d82b14959c6feab2a245167ca