support generic behaviors; implement Mesh resource behavior (#1765)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1765

`hyperactor::behavior!` is expanded to support generic type parameters, allowing for parameterized behaviors.

We use this new functionality to formally define what a mesh controller behavior is, by bundling related types into a trait, so that we can write `Controller<SomeMeshType>`.

Mesh controllers should follow this behavior.

The plan is then to expand this to a set of generic code, tools, and so on, that can be parameterized over different implementations of `Mesh`.
ghstack-source-id: 323322437

Reviewed By: shayne-fletcher, pzhan9

Differential Revision: D86420507

fbshipit-source-id: cef6dc0ed0f2bb500203aa669c3264b057d53260