Option to build oss faster until we remove these (#1880)
Summary:
Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1880

`controller` and `process_allocator` really are no longer needed to develop oss stuff, but we have to remove important testing, etc. For now, I just to be able to compile locally.

This is a 3x build speed improvement because each of these binaries uses different flags for the underlying crates and causes recompiles.
ghstack-source-id: 323494573
exported-using-ghexport

Reviewed By: colin2328

Differential Revision: D86880713

fbshipit-source-id: c43089bcd1140a577d2f1452e3af0d8a826e974e