remote rust_local_mesh (#1947)
Summary:
This diff removes the old way we created tensor engines via rust local mesh. We were no longer testing this in test_remote_functions, etc. but still have a bunch of old examples that depended on it.

The reason to do this now is to lighten what we actually use in the monarch_tensor_worker crate. We need to sever its direct dependence on torch and getting rid of its dead code makes it clearer what we have to do.

Pull Request resolved: https://github.com/meta-pytorch/monarch/pull/1947
ghstack-source-id: 324814292

Reviewed By: mariusae

Differential Revision: D87496994

fbshipit-source-id: 1be151f883bb76ef9fbcac1fc6f605c1a3f285c9